+++ crates/datafusion/src/lib.rs
@@ -39,7 +39,8 @@ use datafusion_expr::{Expr, TableType};
 use datafusion_physical_expr::create_physical_expr;
 
 use hudi_core::config::read::HudiReadConfig::InputPartitions;
-use hudi_core::storage::utils::{empty_options, get_scheme_authority, parse_uri};
+use hudi_core::config::utils::empty_options;
+use hudi_core::storage::utils::{get_scheme_authority, parse_uri};
 use hudi_core::table::Table as HudiTable;
 
 #[derive(Clone, Debug)]
